--- packages/blocks/shuffle/src/assets.d.ts
@@ -1,3 +1,0 @@
-declare module "*.css";
-declare module "*.sass";
-declare module "*.scss";
