+++ packages/blocks/shuffle/src/components/item.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import DragHandleIcon from "@mui/icons-material/DragHandle";
 import DeleteIcon from "@mui/icons-material/Delete";
+import DragHandleIcon from "@mui/icons-material/DragHandle";
 import { Box, IconButton, ListItem, TextField } from "@mui/material";
 import { FunctionComponent } from "react";
 import { Draggable } from "react-beautiful-dnd";
